Resolved race condition in reference counting
Within each thread, a vtkPointLocator is created and its input
is set to the input for the vtkGaussianScalarSplatter. This caused
frequent segmentation faults that I suspect stem from race conditions
in the reference counting increment/decrement operations on the input.

Shallow copying the input works to resolve this issue, but only when
the point locator is not used, i.e., all the points in the input are
processed for all the voxels in the output. This is because there
seems to be another race condition caused by the point locator
modifying some data in the original input despite being passed a
shallow copy of it. Deep copying solves the race condition with the
point locator, but at the cost of copying the data set once for each
thread.

I added a macro called USE_LOCATOR that can be commented out to use
shallow copying or uncommented to use deep copying with the
locator. There is likely an alternative solution that is better than
either of these.