The MADAI logo will now appear only once
You can connect to as many servers as you'd like or load as many state
files as you would like as many times as you like and yet
there will be only one logo.

It does this by removing any MADAI logo entry in the state file before
the XML is processed.